更新變數設置

+++ 20240710/Q1(cincout).cpp
@@ -7,7 +7,7 @@ int main() {
     string name;
     char department;
     int test1, test2;
-    double average;
+    float average;
 
     cout << "Please input your name... ";
     cin >> name;
